+++ vite.config.js
@@ -5,6 +5,12 @@ export default defineConfig({
     base: "/",
     build: {
         outDir: "../dist",
-        emptyOutDir: true
+        emptyOutDir: true,
+        rollupOptions: {
+            input: {
+                main: new URL("./src/index.html", import.meta.url).pathname,
+                impressum: new URL("./src/impressum.html", import.meta.url).pathname
+            }
+        }
     }
 });